+++ example/http/server/flex/http_server_flex.cpp
@@ -263,7 +263,7 @@ class session
                         self_.derived().shared_from_this(),
                         std::placeholders::_1,
                         std::placeholders::_2,
-                        ! sp->keep_alive())));
+                        sp->need_eof())));
         }
     };