+++ example/http/server/stackless/http_server_stackless.cpp
@@ -252,7 +252,7 @@ class session
                         self_.shared_from_this(),
                         std::placeholders::_1,
                         std::placeholders::_2,
-                        ! sp->keep_alive())));
+                        sp->need_eof())));
         }
     };